refactor(editor): extract corpus mock helper for e2e specs
The mockCorpusApi/loadCorpus/loadScenario functions in
edit-test-loop.spec.js are about to be reused by a sister spec that
exercises the form-driven Machine panel flow. Move them into
`frontend/e2e/helpers-corpus.js` so the two specs share one source
of truth and don't drift.

No behavior change; edit-test-loop.spec.js still passes verbatim
against the helper import.
